Avoid rendering content into DOM until popper element created
Fixes #341 and more adequately addresses concerns from #340, #330
without requiring the user to check `isShown`. Bug was introduced
by me in b83580c43689d922cb53e002479781332f9be4ef, and broke a
promise about how enableLazyRendering in 2.x would essentially be
the default and behave the same way.

However, the fastboot tests established that the content would be
rendered into the DOM in fastboot mode, so this preserves that
behavior.