Fix generation of partial sourcemaps
Copied approach used by @jasonmit in ember-cli-showdown. This strips the sourceMappingURL from the popper.js and tooltip.js dependencies, because it seems something cuts the vendor.js.map generation short if there's an existing sourceMappingURL in the source already, so this was causing only Popper.js to have source mapping in production.