Remove CI for Ember 2.18 LTS
Some transitive dependency update is causing the
ember-try 2.18 scenario to fail CI, with a error
related to setTimeout not being defined while trying
to compile a helper in ember-in-element-polyfill.

Not really interested in investigating this further,
so disabling it CI for now.

Ember 2.18 LTS will continue to be supported for 3.x,
but we just won't be testing against it (unless the
issue resolves itself.) It's a three year-old release,
and we're not planning on making any broad sweeping
changes to the 3.x series, so I feel okay about not
having automated tests for 2.18.